# update qcprot to cython.floating types