--- a/testsuite/MDAnalysisTests/analysis/test_rms.py
+++ b/testsuite/MDAnalysisTests/analysis/test_rms.py
@@ -88,11 +88,15 @@ class Testrmsd(object):
                         center=False)
         assert_almost_equal(rmsd, 1.0)
 
-    def test_superposition(self, a, b, u):
+    @pytest.mark.parametrize('dtype', [np.float32, np.float64])
+    def test_superposition(self, a, b, u, dtype):
         bb = u.atoms.select_atoms('backbone')
         a = bb.positions.copy()
         u.trajectory[-1]
         b = bb.positions.copy()
+
+        a, b = a.astype(dtype), b.astype(dtype)
+
         rmsd = rms.rmsd(a, b, superposition=True)
         assert_almost_equal(rmsd, 6.820321761927005)