use memoryviews throughout for potential performance

allows either float32 or float64 inputs

Related #3927 

Changes made in this Pull Request:
 - qcprot functions take either float32 or float64 inputs
 - modernised qcprot to use memoryviews


PR Checklist
------------
 - [x] Tests?
 - [ ] Docs?
 - [x] CHANGELOG updated?
 - [ ] Issue raised/referenced?

## Developers certificate of origin
- [ ] I certify that this contribution is covered by the LGPLv2.1+ license as defined in our [LICENSE](https://github.com/MDAnalysis/mdanalysis/blob/develop/LICENSE) and adheres to the [**Developer Certificate of Origin**](https://developercertificate.org/).


<!-- readthedocs-preview mdanalysis start -->
----
:books: Documentation preview :books:: https://mdanalysis--4273.org.readthedocs.build/en/4273/

<!-- readthedocs-preview mdanalysis end -->