@richardjgowers 2023-09-02
Still need to do some tests here.  `rms.RMSD` looks like it's converting/copying everything up to float64, but I imagine that not forcing this would be some sort of precision-regression.  Instead I can probably write some tests that using `lib.qcprot` directly with floats/doubles work as expected and we can have some future discussion about precision and dtypes in the future.