@richardjgowers 2023-09-03
@orbeckst benchmarks
```python
from MDAnalysisTests.datafiles import PSF, DCD
import MDAnalysis as mda
from MDAnalysis.lib import qcprot
import numpy as np

u = mda.Universe(PSF, DCD)

prot = u.select_atoms('protein')

w = (prot.masses / np.mean(prot.masses)).astype(np.float64)

ref = prot.positions.astype(np.float64)
u.trajectory[1]
conf = prot.positions.astype(np.float64)

rot = np.zeros(9, dtype=np.float64)
A = np.zeros(9)

N = len(prot)
```
then:
```python
%%timeit

qcprot.CalcRMSDRotationalMatrix(ref, conf, len(prot), rot, w)
```
old code:
`10.1 µs ± 46.4 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)`
new code:
`9.97 µs ± 29.1 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)`

The number fluctuate a bit, I think there's no performance gain/loss.

I was able to reproduce the memoryview regression that @tylerjereddy flagged up, but I think it's an order of magnitude smaller than the numbers measured here, so it won't show up.  It's something to keep in mind if we're writing code that has a lot of function calls (there's 3 here).

re: precision, yes I had to make the qcp function use doubles throughout to keep it working.  The `InnerProduct` can use either float or double now.  Strangely these take the same amount of time, which means the compiler isn't able to do anything clever here (where it ought to be) so probably some future digging.  This does some ground work at least.